refactor function to utils
